// src/app/enums/workflow-status.enum.ts
// State: the particular condition that someone or something is in at a specific time.
// Status: the situation at a particular time during a *process.*

export enum WorkflowStatusEnum {

  UNKNOWN = 'UNKNOWN',

  NEW_UNPUBLISHED_LOCKED = 'NEW_UNPUBLISHED_LOCKED',
  NEW_UNPUBLISHED_UNLOCKED = 'NEW_UNPUBLISHED_UNLOCKED',
  NEW_SUBMITTED_WITH_WF_SCHEDULED = 'NEW_SUBMITTED_WITH_WF_SCHEDULED',
  NEW_SUBMITTED_WITH_WF_SCHEDULED_LOCKED = 'NEW_SUBMITTED_WITH_WF_SCHEDULED_LOCKED',
  NEW_SUBMITTED_WITH_WF_UNSCHEDULED = 'NEW_SUBMITTED_WITH_WF_UNSCHEDULED',
  NEW_SUBMITTED_WITH_WF_UNSCHEDULED_LOCKED = 'NEW_SUBMITTED_WITH_WF_UNSCHEDULED_LOCKED',
  NEW_SUBMITTED_NO_WF_SCHEDULED = 'NEW_SUBMITTED_NO_WF_SCHEDULED',
  NEW_SUBMITTED_NO_WF_SCHEDULED_LOCKED = 'NEW_SUBMITTED_NO_WF_SCHEDULED_LOCKED',
  NEW_SUBMITTED_NO_WF_UNSCHEDULED = 'NEW_SUBMITTED_NO_WF_UNSCHEDULED', // waiting in the queue, will be published any moment
  NEW_PUBLISHING_FAILED = 'NEW_PUBLISHING_FAILED',
  NEW_DELETED = 'NEW_DELETED',

  EXISTING_UNEDITED_LOCKED = 'EXISTING_UNEDITED_LOCKED',
  EXISTING_UNEDITED_UNLOCKED = 'EXISTING_UNEDITED_UNLOCKED',
  EXISTING_EDITED_LOCKED = 'EXISTING_EDITED_LOCKED',
  EXISTING_EDITED_UNLOCKED = 'EXISTING_EDITED_UNLOCKED',
  EXISTING_SUBMITTED_WITH_WF_SCHEDULED = 'EXISTING_SUBMITTED_WITH_WF_SCHEDULED',
  EXISTING_SUBMITTED_WITH_WF_SCHEDULED_LOCKED = 'EXISTING_SUBMITTED_WITH_WF_SCHEDULED_LOCKED',
  EXISTING_SUBMITTED_WITH_WF_UNSCHEDULED = 'EXISTING_SUBMITTED_WITH_WF_UNSCHEDULED',
  EXISTING_SUBMITTED_WITH_WF_UNSCHEDULED_LOCKED = 'EXISTING_SUBMITTED_WITH_WF_UNSCHEDULED_LOCKED',
  EXISTING_SUBMITTED_NO_WF_SCHEDULED = 'EXISTING_SUBMITTED_NO_WF_SCHEDULED',
  EXISTING_SUBMITTED_NO_WF_SCHEDULED_LOCKED = 'EXISTING_SUBMITTED_NO_WF_SCHEDULED_LOCKED',
  EXISTING_SUBMITTED_NO_WF_UNSCHEDULED = 'EXISTING_SUBMITTED_NO_WF_UNSCHEDULED',
  EXISTING_PUBLISHING_FAILED = 'EXISTING_PUBLISHING_FAILED',
  EXISTING_DELETED = 'EXISTING_DELETED'

}